got save button and edit button to work with sidebar updating

+++ src/components/workspacePlaceholder.js
@@ -1,14 +1,19 @@
 import "./Notes.css";
-import DateTime from "./DateTime";
+import DateTime, { monthNames } from "./DateTime";
 import Quill from "./Quill";
 import { useState, useEffect } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
 
-const WSPlaceHolder = ({ selectedNote, updateTitle, setSelectedNote, updateIsSaved }) => {
+const WSPlaceHolder = ({
+	selectedNote,
+	title,
+	setTitle,
+	setSelectedNote,
+	updateIsSaved,
+	saveTitle,
+}) => {
 	const navigate = useNavigate();
 	const location = useLocation();
-
-	const [title, setTitle] = useState(selectedNote ? selectedNote.title : "");
 
 	const [date, setDate] = useState(null);
 	const [text, setText] = useState(null);
@@ -18,54 +23,71 @@ const WSPlaceHolder = ({ selectedNote, updateTitle, setSelectedNote, updateIsSav
 
 	const [isSaved, setIsSaved] = useState(false);
 
+	const [isEdit, setIsEdit] = useState(false);
+
 	const handleSave = () => {
-		const newData = {
-			title: title,
-			date: date,
-			text: text,
-		};
-
-		const saveData = [...data, newData];
-		setData(saveData);
-		localStorage.setItem("data", JSON.stringify(data));
-
-		const newPath = location.pathname.replace("/edit", "");
-		navigate(newPath, { replace: true });
-
-		setIsReadOnly(true);
-
-		setTitle(title);
-		updateTitle(title);
-		localStorage.setItem("title", title);
-
-		setIsSaved(true);
-		updateIsSaved(true);
-
-		const textArray = text.split(" ");
-		let previewText = "";
-		for (let i = 0; i < Math.min(textArray.length, 12); i++) {
-			previewText += textArray[i] + " ";
+		if (title) {
+			setTitle(title);
 		}
-		previewText = previewText.trim();
+		if (!isEdit) {
+			saveTitle();
 
-		previewText = previewText.length > 80 ? previewText.substring(0, 80) + "..." : previewText;
+			const newData = {
+				title: title,
+				date: date,
+				text: text,
+			};
 
-		setSelectedNote({ ...selectedNote, title, previewText });
+			const saveData = [...data, newData];
+			setData(saveData);
+			localStorage.setItem("data", JSON.stringify(data));
 
-		window.onbeforeunload = function () {
-			localStorage.clear();
-		};
+			const newPath = location.pathname.replace("/edit", "");
+			navigate(newPath, { replace: true });
+
+			setIsReadOnly(true);
+			setTitle(title);
+
+			setIsSaved(true);
+			updateIsSaved(true);
+
+			const textArray = text.split(" ");
+			let previewText = "";
+			for (let i = 0; i < Math.min(textArray.length, 12); i++) {
+				previewText += textArray[i] + " ";
+			}
+			previewText = previewText.trim();
+
+			previewText =
+				previewText.length > 80 ? previewText.substring(0, 80) + "..." : previewText;
+
+			const dateObject = new Date(date);
+			dateObject.setMinutes(dateObject.getMinutes() - 420);
+			const saveDate = `${
+				monthNames[dateObject.getMonth()]
+			} ${dateObject.getDate()}, ${dateObject.getFullYear()} at ${dateObject.toLocaleTimeString(
+				[],
+				{ hour: "2-digit", minute: "2-digit", hour12: false }
+			)}`;
+
+			setSelectedNote({ ...selectedNote, title, previewText, saveDate });
+
+			window.onbeforeunload = function () {
+				localStorage.clear();
+			};
+
+			setIsEdit(true);
+		} else {
+			navigate(location.pathname + "/edit", { replace: true });
+			setIsReadOnly(!isReadOnly);
+			setIsSaved(false);
+			setIsEdit(false);
+		}
 	};
 
 	useEffect(() => {
 		setDate(new Date().toISOString().substring(0, 16));
 	}, []);
-
-	useEffect(() => {
-		localStorage.setItem("title", title);
-		localStorage.setItem("date", JSON.stringify(date));
-		localStorage.setItem("text", JSON.stringify(text));
-	}, [title, date, text]);
 
 	useEffect(() => {
 		localStorage.setItem("data", JSON.stringify(data));
@@ -76,15 +98,20 @@ const WSPlaceHolder = ({ selectedNote, updateTitle, setSelectedNote, updateIsSav
 			<div className="workBox">
 				<input
 					className="inputBox"
-					value={(selectedNote && selectedNote.title) || ""}
-					onChange={(e) => {
-						setTitle(e.target.value);
-						updateTitle(e.target.value);
-					}}
-					disabled={isReadOnly}></input>
+					value={title || selectedNote.title || ""}
+					onChange={(e) => setTitle(e.target.value)}
+					disabled={isReadOnly || isEdit}
+				/>
+
 				<DateTime onChange={(value) => setDate(value)} readOnly={isReadOnly} />
-				<button className="saveButton" onClick={handleSave}>
-					Save
+
+				<button
+					className="saveButton"
+					onClick={() => {
+						saveTitle();
+						handleSave();
+					}}>
+					{isEdit ? "Edit" : "Save"}
 				</button>
 				<button className="delButton">Delete</button>
 			</div>
